+++ lib/src/license.g.dart
@@ -10,14 +10,17 @@ part of pana.license;
 // Generator: JsonSerializableGenerator
 // **************************************************************************
 
-License _$LicenseFromJson(Map<String, dynamic> json) =>
-    new License(json['name'] as String, json['version'] as String);
+LicenseFile _$LicenseFileFromJson(Map<String, dynamic> json) =>
+    new LicenseFile(json['path'] as String, json['name'] as String,
+        version: json['version'] as String);
 
-abstract class _$LicenseSerializerMixin {
+abstract class _$LicenseFileSerializerMixin {
+  String get path;
   String get name;
   String get version;
   Map<String, dynamic> toJson() {
     var val = <String, dynamic>{
+      'path': path,
       'name': name,
     };