+++ test/end2end/pub_server_data.dart
@@ -353,7 +353,10 @@ final _data = {
   },
   'platform': {'worksEverywhere': true, 'reason': 'All libraries agree'},
   "licenses": [
-    {"name": "BSD"},
+    {
+      "path": "LICENSE",
+      "name": "BSD",
+    },
   ],
   'fitness': {'magnitude': 455.0, 'shortcoming': 1.0},
 };